feat: updated nuxt examples
This PR includes the addition of a health check API endpoint directly in the Nuxt example. This API endpoint can be used as an alternative to the default health check on port 80.

Beside the API feature I also included a Nitro example. Nitro can be run seperately from Nuxt. Allowing you to basically run Nuxt 'headless'. This extension of the existing example will go well with developers using Nuxt within the Coolify community.

This PR is complementary to my PR on the main Coolify documentation PR.